Changed Mesh to ResourceMesh
